+++ packages/ra-core/src/CoreAdmin.tsx
@@ -1,9 +1,4 @@
-import React, {
-    createElement,
-    Component,
-    ReactNode,
-    ComponentType,
-} from 'react';
+import React, { createElement, Component, ComponentType, SFC } from 'react';
 import PropTypes from 'prop-types';
 import { Provider } from 'react-redux';
 import { History } from 'history';
@@ -15,37 +10,60 @@ import withContext from 'recompose/withContext';
 import createAdminStore from './createAdminStore';
 import TranslationProvider from './i18n/TranslationProvider';
 import CoreAdminRouter from './CoreAdminRouter';
-import { AuthProvider, I18nProvider, DataProvider } from './types';
+import {
+    AuthProvider,
+    I18nProvider,
+    DataProvider,
+    TitleComponent,
+    LoginComponent,
+    LayoutComponent,
+    LayoutProps,
+    AdminChildren,
+    CatchAllComponent,
+    CustomRoutes,
+    DashboardComponent,
+} from './types';
 
 export type ChildrenFunction = () => ComponentType[];
 
-const DefaultLayout = ({ children }) => children;
+const DefaultLayout: SFC<LayoutProps> = ({ children }) => <>{children}</>;
 
-interface Props {
-    appLayout: ComponentType;
-    authProvider: AuthProvider;
-    children: ReactNode | ChildrenFunction;
-    catchAll: ComponentType;
-    customSagas: any[];
-    customReducers: object;
-    customRoutes: any[];
-    dashboard: ComponentType;
+export interface AdminProps {
+    appLayout: LayoutComponent;
+    authProvider?: AuthProvider;
+    children?: AdminChildren;
+    catchAll: CatchAllComponent;
+    customSagas?: any[];
+    customReducers?: object;
+    customRoutes?: CustomRoutes;
+    dashboard?: DashboardComponent;
     dataProvider: DataProvider;
     history: History;
-    i18nProvider: I18nProvider;
-    initialState: object;
+    i18nProvider?: I18nProvider;
+    initialState?: object;
     loading: ComponentType;
-    locale: string;
-    loginPage: ComponentType | boolean;
-    logoutButton: ComponentType;
-    menu: ComponentType;
-    theme: object;
-    title: ReactNode;
+    locale?: string;
+    loginPage: LoginComponent | boolean;
+    logoutButton?: ComponentType;
+    menu?: ComponentType;
+    theme?: object;
+    title?: TitleComponent;
 }
 
-class CoreAdmin extends Component<Props> {
+interface AdminContext {
+    authProvider: AuthProvider;
+}
+
+class CoreAdminBase extends Component<AdminProps> {
     static contextTypes = {
         store: PropTypes.object,
+    };
+
+    static defaultProps: Partial<AdminProps> = {
+        catchAll: () => null,
+        appLayout: DefaultLayout,
+        loading: () => null,
+        loginPage: false,
     };
 
     reduxIsAlreadyInitialized = false;
@@ -72,7 +90,7 @@ React-admin requires a valid dataProvider function to work.`);
 
     renderCore() {
         const {
-            appLayout = DefaultLayout,
+            appLayout,
             authProvider,
             children,
             customRoutes = [],
@@ -98,12 +116,12 @@ React-admin requires a valid dataProvider function to work.`);
             <TranslationProvider>
                 <ConnectedRouter history={this.history}>
                     <Switch>
-                        {loginPage !== false ? (
+                        {loginPage !== false && loginPage !== true ? (
                             <Route
                                 exact
                                 path="/login"
                                 render={props =>
-                                    createElement(loginPage as ComponentType, {
+                                    createElement(loginPage, {
                                         ...props,
                                         title,
                                         theme,
@@ -120,7 +138,6 @@ React-admin requires a valid dataProvider function to work.`);
                                     customRoutes={customRoutes}
                                     dashboard={dashboard}
                                     loading={loading}
-                                    loginPage={loginPage}
                                     logout={logout}
                                     menu={menu}
                                     theme={theme}
@@ -169,9 +186,11 @@ React-admin requires a valid dataProvider function to work.`);
     }
 }
 
-export default withContext(
+const CoreAdmin = withContext<AdminContext, AdminProps>(
     {
         authProvider: PropTypes.func,
     },
     ({ authProvider }) => ({ authProvider })
-)(CoreAdmin);
+)(CoreAdminBase) as ComponentType<AdminProps>;
+
+export default CoreAdmin;