+++ packages/ra-core/src/form/addField.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import FormField from './FormField';
-
-export default (BaseComponent, fieldProps = {}) => {
-    const WithFormField = props => (
-        <FormField component={BaseComponent} {...fieldProps} {...props} />
-    );
-    return WithFormField;
-};