+++ packages/ra-core/src/i18n/translate.tsx
@@ -9,7 +9,7 @@ import {
 /**
  * Higher-Order Component for getting access to the `translate` function in props.
  *
- * Requires that the app is decorated by the <TranslationPRovider> to inject
+ * Requires that the app is decorated by the <TranslationProvider> to inject
  * the translation dictionaries and function in the context.
  *
  * @example